Remove Sketch class. Add new skeleton code for work on renderer.
We were only using on instance of the `Sketch` object so there was no
need to have a separate class when a module would suffice. Since we
will start work on the renderer now, we also setup a skeleton p5
sketch, a super minimal `Shape` class, and the `point()` function. For
now, as shapes are created, we add them to a temporary
`_tmp_on_screen_shapes` list and draw all of them on each `on_draw()`.