FIX: Windows shader compilation exception
On windows (with OpenGL 3.3 with an intel card) produced a "No Errors"
log message on a successful shader compilation. The shader compilation
step raised an exception when the this log was non-empty and hence,
was raising an exception even on a successful compilation.

THIS IS A TEMPORARY WORKAROUND until I find a better way to handle
this.