Fixed bug #108 .
Fixes #108 .

In the previous code, the variables ` self.redraw`  and ` self.looping`  were assigned after the ` setup()`  function is called. This prevented the ` loop()`, ` no_loop()`  and ` redraw()` to function correctly inside the ` setup()`  function.

I have fixed the issue by defining the variables self.looping and self.redraw as ` None` in the beginning and assign them after setup() function is called if they are not assigned inside the setup() function (line 145 and 147)