Fix exceptions shown when closing canvas
`p5.exit()` documentation mentions that it forwards args
to the builtin `exit()`, so do that instead of calling
the non-existing `p5.core.p5.exit()` function. (Or prevent
a recursive call that by exit ended up calling a
non-existant function.

Do not call `p5.exit()` in on_close: exit() will throw a SystemExit
exception which will be cought by vispy, reported and ignored.
Since the windows has gone away we only need to quit the
running application for a normal exit.